Fixed any type errors, and created a new token component while hiding secret credentials
