+++ ui/.gitignore
@@ -23,3 +23,5 @@ yarn-debug.log*
 yarn-error.log*
 
 /yarn.lock
+
+.env
