+++ ui/src/components/AlbumPage.tsx
@@ -1,18 +1,15 @@
 import React, { useEffect, useState } from "react";
 import { useParams } from "react-router";
 import NavBar from "./NavBar";
-import { Track, Album } from "../types";
+import { Track, Album } from "../types/types";
 import ExplicitIcon from "@mui/icons-material/Explicit";
 import PlayCircleRoundedIcon from "@mui/icons-material/PlayCircleRounded";
 import { green, red } from "@mui/material/colors";
 import FavoriteBorderOutlinedIcon from "@mui/icons-material/FavoriteBorderOutlined";
-
-const client_id = "c01ff52afa434ea490f0081740544b71";
-const client_secret = "5b0bb494f9964251be58dcb8f1eba45e";
+import Auth from "./Auth";
 
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 export default function AlbumPage() {
-  const [accessToken, setAccessToken] = useState("");
   const { id } = useParams();
   const [albumName, setAlbumName] = useState("");
   const [albumTracks, setAlbumTracks] = useState<Track[]>([]);
@@ -20,23 +17,7 @@ export default function AlbumPage() {
   const [albumArtist, setAlbumArtist] = useState("");
   const [albumDate, setAlbumDate] = useState("");
   const [total, setTotal] = useState("");
-
-  useEffect(() => {
-    //API access token
-    const authParameters = {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/x-www-form-urlencoded",
-      },
-      body: `grant_type=client_credentials&client_id=${client_id}&client_secret=${client_secret}`,
-    };
-    fetch("https://accounts.spotify.com/api/token", authParameters)
-      .then((result) => result.json())
-      .then((data) => {
-        console.log(data);
-        setAccessToken(data.access_token);
-      });
-  }, []);
+  const accessToken = Auth();
 
   // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
   const getAlbum = () => {
