+++ ui/src/components/Login.tsx
@@ -1,7 +1,6 @@
 import React from "react";
 
-const AUTH_URL =
-  "https://accounts.spotify.com/authorize?client_id=c01ff52afa434ea490f0081740544b71&response_type=code&redirect_uri=http://localhost:3000/dashboard&scope=streaming%20user-read-email%20user-read-private%20user-library-read%20user-library-modify%20user-read-playback-state%20user-modify-playback-state%20playlist-read-private%20playlist-read-collaborative%20user-top-read";
+const AUTH_URL = `https://accounts.spotify.com/authorize?client_id=${process.env.SPOTIFY_CLIENT_ID}&response_type=code&redirect_uri=${process.env.REDIRECT_URI}&scope=streaming%20user-read-email%20user-read-private%20user-library-read%20user-library-modify%20user-read-playback-state%20user-modify-playback-state%20playlist-read-private%20playlist-read-collaborative%20user-top-read`;
 // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 export default function Login() {
   return (
