+++ ui/src/index.tsx
@@ -1,11 +1,16 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
-import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 import "./styles/tailwind.css";
 import { StyledEngineProvider } from "@mui/material";
-import AppRouter from "./router/AppRouter";
+import { BrowserRouter, Route, Routes } from "react-router-dom";
+import Login from "./components/Login";
+import Dashboard from "./components/Dashboard";
+import SearchBar from "./components/SearchBar";
+import AlbumPage from "./components/AlbumPage";
+import PlaylistPage from "./components/PlaylistPage";
+import Library from "./components/Library";
 
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
@@ -13,7 +18,16 @@ const root = ReactDOM.createRoot(
 root.render(
   <React.StrictMode>
     <StyledEngineProvider injectFirst>
-      <AppRouter />
+      <BrowserRouter>
+        <Routes>
+          <Route path="/" element={<Login />} />
+          <Route path="/dashboard" element={<Dashboard />} />
+          <Route path="/search" element={<SearchBar />} />
+          <Route path="/album/:id" element={<AlbumPage />} />
+          <Route path="/playlist/:id" element={<PlaylistPage />} />
+          <Route path="/library" element={<Library />} />
+        </Routes>
+      </BrowserRouter>
     </StyledEngineProvider>
   </React.StrictMode>
 );
