+++ ui/tailwind.config.js
@@ -2,9 +2,7 @@ module.exports = {
     content: [
         "./src/**/*.{js,jsx,ts,tsx}",
     ],
-    corePlugins: {
-        preflight: false,
-    },
+
     important: '#root',
     theme: {
         extend: {
