+++ ui/tailwind/tailwind.js
@@ -1,11 +1,9 @@
 module.exports = {
-	theme: {
-		extend: {
-
-		}
-	},
-	future: {
-		removeDeprecatedGapUtilities: true,
-		purgeLayersByDefault: true,
-	},
+    theme: {
+        extend: {}
+    },
+    future: {
+        removeDeprecatedGapUtilities: true,
+        purgeLayersByDefault: true,
+    },
 };
